truthlens-anti-deepfake/ai-model-service/requirements.txt
flask
werkzeug
opencv-python-headless
numpy
tensorflow-cpu
